backend/requirements.txt
absl-py==0.5.0
astor==0.7.1
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
itsdangerous==0.24
Jinja2==2.10
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==3.0
MarkupSafe==1.0
numpy==1.14.5
Pillow==5.2.0
protobuf==3.6.1
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
Werkzeug==0.14.1
